refactor: sidebar breaking down to components
